media: mtk-mdp: porting code from 2701 and 8173
Add more color format support.
Test: pass

CR-Id: AUTO00003633

+++ drivers/media/platform/mtk-mdp/mtk_mdp_vpu.h
@@ -40,7 +40,7 @@
  */
 struct mtk_mdp_vpu {
 	struct platform_device	*pdev;
-	uint32_t		inst_addr;
+	uint64_t		inst_addr;
 	int32_t			failure;
 	struct mdp_process_vsi	*vsi;
 };
